requirements.txt
Pillow
autograd
cvxpy
matplotlib
numpy
torch
torchvision
tqdm
opencv-python